pyproject.toml
[tool.poetry]
name = "icarus-x"
version = "1.0.0"
description = "ICARUS-X: Unified AI-Powered Pentesting Framework"
authors = ["PenTrix Security"]
readme = "README.md"
packages = [{include = "icarus_x"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
httpx = "^0.26.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
aiodns = "^3.1.0"
aiofiles = "^23.2.0"
aiohttp = "^3.9.0"
sqlmodel = "^0.0.14"
aiosqlite = "^0.19.0"
jinja2 = "^3.1.0"
orjson = "^3.9.0"
python-whois = "^0.9.0"
google-generativeai = "^0.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
black = "^24.1.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
icx = "icarus:app"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
